сделал генерацию картинок на главной
